Fixed #4; set width of textfields to 280.dp

+++ app/src/main/java/com/wcsm/shopperrotas/data/model/RideEstimateResponse.kt
@@ -3,7 +3,7 @@ package com.wcsm.shopperrotas.data.model
 data class RideEstimateResponse(
     val origin: Location,
     val destination: Location,
-    val distance: Int,
+    val distance: Double,
     val duration: String,
     val options: List<RideOption>,
     val routeResponse: Any
